refactor: cleaned up color imports
