+++ bot/cogs/coffee.py
@@ -4,7 +4,7 @@ from discord import Member, Embed
 from discord.ext import commands
 from discord.ext.commands import Context
 from .inputs import cl, cf, chill, cfe, ur
-from .utils import COLOR
+from .utils.colo import COLOR
 
 
 class Coffee(commands.Cog):
