+++ bot/cogs/comic.py
@@ -6,7 +6,7 @@ import aiohttp
 from discord import Embed
 from discord.ext.commands import Context
 from discord.ext import commands
-from .utils import COLOR
+from .utils.colo import COLOR
 
 
 class Comics(commands.Cog):
