+++ bot/cogs/error.py
@@ -2,7 +2,7 @@ from discord import Embed
 from discord.ext import commands
 from discord.ext.commands import Context
 
-from .utils import COLOR
+from .utils.colo import COLOR
 
 
 class Errors(commands.Cog):
